fixed position and sqrtf

+++ src/intersections/hit_sp.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   hit_sp.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jeepark <jeepark@student.42.fr>            +#+  +:+       +#+        */
+/*   By: cgosseli <cgosseli@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/21 14:48:49 by jeepark           #+#    #+#             */
-/*   Updated: 2022/09/25 16:40:03 by jeepark          ###   ########.fr       */
+/*   Updated: 2022/10/01 16:00:59 by cgosseli         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,8 +32,8 @@ t_hit_point hit_sp(t_ray *ray, t_object *sp)
         return (hit);
 	}
 	hit.status = 1;
-	t1 = (- hit.b - sqrtf(discr)) / (2.0 * hit.a);
-	t2 = (- hit.b + sqrtf(discr)) / (2.0 * hit.a);
+	t1 = (- hit.b - sqrt(discr)) / (2.0 * hit.a);
+	t2 = (- hit.b + sqrt(discr)) / (2.0 * hit.a);
 	if (t1 < t2)
 		hit.root = t1;
 	else
